# docs/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Musicfy API",
...
    "contact": {
      "name": "Musicfy Backend"
    },
    "version": "1.0.0"
  },
  "servers": [
    {
      "url": "http://localhost:8000",
      "description": "Local"
    }
  ],
  "paths": {
...
    "/v1/lyrics": {
      "post": {
        "tags": [
          "Текст песни"
        ],
        "summary": "Сгенерировать текст",
        "operationId": "generate_lyrics_v1_lyrics_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/GenerateLyricsRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/LyricsDraftResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ]
      }
    },
    "/v1/lyrics/{draft_id}": {
      "get": {
        "tags": [
          "Текст песни"
        ],
        "summary": "Получить текст",
        "operationId": "get_lyrics_v1_lyrics__draft_id__get",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "draft_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Draft Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/LyricsDraftResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "patch": {
        "tags": [
          "Текст песни"
        ],
        "summary": "Отредактировать текст",
        "operationId": "update_lyrics_v1_lyrics__draft_id__patch",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "draft_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Draft Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateLyricsRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/LyricsDraftResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/v1/jobs/{job_id}": {
      "get": {
        "tags": [
          "Задачи"
        ],
        "summary": "Статус задачи",
        "operationId": "get_job_v1_jobs__job_id__get",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "job_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Job Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/JobStatusResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/v1/billing/balance": {
      "get": {
        "tags": [
          "Биллинг"
        ],
        "summary": "Баланс генераций",
        "operationId": "balance_v1_billing_balance_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/BalanceResponse"
                }
              }
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ]
      }
    },
...
    "/v1/billing/purchases/verify": {
      "post": {
        "tags": [
          "Биллинг"
        ],
        "summary": "Проверить покупку",
        "operationId": "verify_purchase_v1_billing_purchases_verify_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/VerifyPurchaseRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApplyResultResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ]
      }
    },
    "/v1/billing/restore": {
      "post": {
        "tags": [
          "Биллинг"
        ],
        "summary": "Restore purchases",
        "operationId": "restore_v1_billing_restore_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RestoreRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/ApplyResultResponse"
                  },
                  "type": "array",
                  "title": "Response Restore V1 Billing Restore Post"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ]
      }
    },
    "/v1/billing/ledger": {
      "get": {
        "tags": [
          "Биллинг"
        ],
        "summary": "Журнал кредитов",
        "operationId": "ledger_v1_billing_ledger_get",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "default": 100,
              "title": "Limit"
            }
          },
          {
            "name": "offset",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "default": 0,
              "title": "Offset"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/LedgerEntryView"
                  },
                  "title": "Response Ledger V1 Billing Ledger Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "ApplyResultResponse": {
        "properties": {
          "status": {
            "type": "string",
            "title": "Status"
          },
          "deduplicated": {
            "type": "boolean",
            "title": "Deduplicated",
            "default": false
          }
        },
        "type": "object",
        "required": [
          "status"
        ],
        "title": "ApplyResultResponse"
      },
...
      "BalanceResponse": {
        "properties": {
          "balances": {
            "items": {
              "$ref": "#/components/schemas/CategoryBalance"
            },
            "type": "array",
            "title": "Balances"
          }
        },
        "type": "object",
        "required": [
          "balances"
        ],
        "title": "BalanceResponse"
      },
...
      "CategoryBalance": {
        "properties": {
          "category": {
            "type": "string",
            "title": "Category"
          },
          "subscriptionRemaining": {
            "type": "integer",
            "title": "Subscriptionremaining"
          },
          "subscriptionGranted": {
            "type": "integer",
            "title": "Subscriptiongranted"
          },
          "periodEnd": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ],
            "title": "Periodend"
          },
          "purchasedAvailable": {
            "type": "integer",
            "title": "Purchasedavailable"
          }
        },
        "type": "object",
        "required": [
          "category",
          "subscriptionRemaining",
          "subscriptionGranted",
          "periodEnd",
          "purchasedAvailable"
        ],
        "title": "CategoryBalance"
      },
...
      "GenerateLyricsRequest": {
        "properties": {
          "prompt": {
            "type": "string",
            "maxLength": 2000,
            "minLength": 1,
            "title": "Prompt",
            "description": "Тема/идея песни."
          },
          "language": {
            "type": "string",
            "maxLength": 8,
            "title": "Language",
            "description": "Язык (ISO-код).",
            "default": "en"
          },
          "genre": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 64
              },
              {
                "type": "null"
              }
            ],
            "title": "Genre"
          },
          "mood": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 64
              },
              {
                "type": "null"
              }
            ],
            "title": "Mood"
          }
        },
        "type": "object",
        "required": [
          "prompt"
        ],
        "title": "GenerateLyricsRequest",
        "description": "Синхронная генерация текста песни. Результат можно править через PATCH."
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "JobStatusResponse": {
        "properties": {
          "jobId": {
            "type": "string",
            "title": "Jobid"
          },
          "jobType": {
            "type": "string",
            "title": "Jobtype"
          },
          "status": {
            "type": "string",
            "title": "Status"
          },
          "currentStage": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Currentstage"
          },
          "errorCode": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Errorcode"
          },
          "errorMessage": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Errormessage"
          },
          "trackId": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Trackid"
          },
          "pipeline": {
            "items": {
              "$ref": "#/components/schemas/StageView"
            },
            "type": "array",
            "title": "Pipeline"
          },
          "createdAt": {
            "type": "string",
            "format": "date-time",
            "title": "Createdat"
          },
          "updatedAt": {
            "type": "string",
            "format": "date-time",
            "title": "Updatedat"
          }
        },
        "type": "object",
        "required": [
          "jobId",
          "jobType",
          "status",
          "currentStage",
          "errorCode",
          "errorMessage",
          "pipeline",
          "createdAt",
          "updatedAt"
        ],
        "title": "JobStatusResponse"
      },
      "LedgerEntryView": {
        "properties": {
          "kind": {
            "type": "string",
            "title": "Kind"
          },
          "category": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Category"
          },
          "source": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Source"
          },
          "amount": {
            "type": "integer",
            "title": "Amount"
          },
          "reason": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Reason"
          },
          "createdAt": {
            "type": "string",
            "format": "date-time",
            "title": "Createdat"
          }
        },
        "type": "object",
        "required": [
          "kind",
          "category",
          "source",
          "amount",
          "reason",
          "createdAt"
        ],
        "title": "LedgerEntryView"
      },
...
      "LyricsDraftResponse": {
        "properties": {
          "id": {
            "type": "string",
            "title": "Id"
          },
          "content": {
            "type": "string",
            "title": "Content"
          },
          "language": {
            "type": "string",
            "title": "Language"
          },
          "genre": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Genre"
          },
          "mood": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Mood"
          },
          "source": {
            "type": "string",
            "title": "Source"
          },
          "createdAt": {
            "type": "string",
            "format": "date-time",
            "title": "Createdat"
          }
        },
        "type": "object",
        "required": [
          "id",
          "content",
          "language",
          "genre",
          "mood",
          "source",
          "createdAt"
        ],
        "title": "LyricsDraftResponse"
      },
...
      "RestoreRequest": {
        "properties": {
          "signedTransactions": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Signedtransactions",
            "description": "Список JWS-транзакций для восстановления."
          }
        },
        "type": "object",
        "title": "RestoreRequest",
        "description": "Restore purchases: массив подписанных транзакций из StoreKit."
      },
...
      "StageView": {
        "properties": {
          "stage": {
            "type": "string",
            "title": "Stage"
          },
          "status": {
            "type": "string",
            "title": "Status"
          },
          "error": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Error"
          }
        },
        "type": "object",
        "required": [
          "stage",
          "status"
        ],
        "title": "StageView"
      },
...
      "UpdateLyricsRequest": {
        "properties": {
          "content": {
            "type": "string",
            "maxLength": 8000,
            "minLength": 1,
            "title": "Content"
          }
        },
        "type": "object",
        "required": [
          "content"
        ],
        "title": "UpdateLyricsRequest"
      },
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      },
      "VerifyPurchaseRequest": {
        "properties": {
          "signedTransaction": {
            "type": "string",
            "minLength": 1,
            "title": "Signedtransaction",
            "description": "JWS signedTransactionInfo из StoreKit 2 (Transaction)."
          }
        },
        "type": "object",
        "required": [
          "signedTransaction"
        ],
        "title": "VerifyPurchaseRequest",
        "description": "Верификация покупки StoreKit 2."
      },
...
    },
    "securitySchemes": {
      "HTTPBearer": {
        "type": "http",
        "description": "Админ-ключ (ADMIN_API_KEY)",
        "scheme": "bearer"
      }
    }
  },
...
}